Implement yaml for runscript encoding
