+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "submodules/runsolver"]
-	path = submodules/runsolver
+[submodule "external-tools/runsolver"]
+	path = external-tools/runsolver
 	url = git@github.com:daajoe/runsolver.git
