+++ Pipfile
@@ -8,6 +8,9 @@ verify_ssl = true
 [packages]
 lxml = "*"
 jinja2 = "*"
+pyyaml = "*"
+pykwalify = "*"
+dotmap = "*"
 
 [requires]
 python_version = "3"
