--- benchmarks/examples/clasp/runscript-all.xml
@@ -1,50 +1,0 @@
-<runscript output="benchmark/examples/clasp/output">
-
-	<machine name="houat" cpu="8xE5520@2.27GHz" memory="24GB"/>
-	<machine name="zuse" cpu="24x8xE5520@2.27GHz" memory="24GB"/>
-	
-	<config name="seq-generic" template="benchmarks/examples/clasp/templates/seq-generic.sh"/>
-	<system name="clasp" version="1.3.2" measures="benchmarks.examples.clasp.resultparsers.clasp" config="seq-generic">
-		<!-- optionally allows for tagging settings -->
-		<setting name="default"  tag="seq" cmdline="--stats 1"/>
-		<setting name="vsids"    tag="seq" cmdline="--stats --heu=vsids 1"/>
-		<setting name="progress" tag="seq" cmdline="--stats --save-progress 1"/>
-		<setting name="estimate" tag="seq" cmdline="--stats --estimate 1"/>
-		<setting name="recstr"   tag="seq" cmdline="--stats --recursive-str 1"/>
-		<setting name="nodel"    tag="seq" cmdline="--stats --del=no 1"/>
-		<setting name="norest"   tag="seq" cmdline="--stats --restarts=no 1"/>
-	</system>
-
-	<config name="pbs-generic" template="benchmarks/examples/clasp/templates/pbs-generic.sh"/>
-	<system name="claspar" version="2.1.0" measures="benchmarks.examples.clasp.resultparsers.claspar" config="pbs-generic">
-		<setting name="one-as" tag="par one-as" cmdline="--stats 1" ppn="2" procs="1 2 4 8" pbstemplate="benchmarks/examples/clasp/templates/impi.pbs"/>
-		<setting name="all-as" tag="par all-as" cmdline="--stats -q 0" ppn="2" procs="1 2 4 8" pbstemplate="benchmarks/examples/clasp/templates/impi.pbs"/>
-	</system>
-	
-	<seqjob name="seq-generic" timeout="120" runs="1" parallel="8"/>
-	<pbsjob name="pbs-generic" timeout="120" runs="1" script_mode="timeout" walltime="23:59:59"/>
-
-	<benchmark name="seq-suite">
-		<folder path="benchmarks/examples/clasp/instances">
-			<ignore prefix="pigeons"/>
-		</folder>
-		<files path="benchmarks/examples/clasp/instances">
-			<add file="pigeons/pigeonhole10-unsat.lp"/>
-			<add file="pigeons/pigeonhole11-unsat.lp"/>
-		</files>
-	</benchmark>
-	<benchmark name="pbs-suite">
-		<folder path="benchmarks/examples/clasp/instances"/>
-	</benchmark>
-
-	<project name="clasp-big" job="seq-generic">
-		<runtag machine="houat" benchmark="seq-suite" tag="seq"/>
-	</project>
-	<project name="claspar-all-as" job="pbs-generic">
-		<runtag machine="houat" benchmark="pbs-suite" tag="par one-as"/>
-	</project>
-	<project name="claspar-one-as" job="pbs-generic">
-		<runspec machine="zuse" benchmark="pbs-suite" system="claspar" version="2.1.0" setting="one-as"/>
-	</project>
-
-</runscript>
