+++ src/benchmarktool/runscript/runscript.py
@@ -16,6 +16,10 @@ from jinja2 import Template
 # pylint: disable-msg=W0611
 import benchmarktool.config  # @UnusedImport
 
+from .jobs.condor import CondorJob
+from .jobs.pbs import PbsJob
+from .jobs.sequential import SeqJob
+
 
 # Taken from SO:312443
 def chunks(l, n):
@@ -79,7 +83,7 @@ class System(Sortable):
     together with a set of settings.  
     """
 
-    def __init__(self, name, version, measures, order):
+    def __init__(self, name, version, measures, order, **kwargs):
         """
         Initializes a system. Name and version of the system 
         are used to uniquely identify a system.   
@@ -144,7 +148,7 @@ class Setting(Sortable):
     that can be passed to the system. Additionally, settings can be tagged. 
     """
 
-    def __init__(self, name, cmdline, tag, order, procs, ppn, pbstemplate, attr):
+    def __init__(self, name, cmdline, tag, order, **kwargs):
         """
         Initializes a system.
         
@@ -158,16 +162,14 @@ class Setting(Sortable):
         procs       - Number of processes used by the solver (pbs only)
         ppn         - Processes per node (pbs only)
         pbstemplate - Path to pbs-template file (pbs only, related to mpi-version)
-        attr        - A dictionary of additional optional attributes.  
         """
         self.name = name
         self.cmdline = escape(cmdline)
         self.tag = tag
         self.order = order
-        self.procs = procs
-        self.ppn = ppn
-        self.pbstemplate = pbstemplate
-        self.attr = attr
+        self.procs = kwargs.get("procs")
+        self.ppn = kwargs.get("ppn")
+        self.pbstemplate = kwargs.get("pbstemplate")
 
     def toXml(self, out, indent):
         """
@@ -200,683 +202,6 @@ class Setting(Sortable):
         Compares two settings using their names. 
         """
         return cmp(self.name, setting.name)
-
-
-class Job(Sortable):
-    """
-    Base class for all jobs. 
-    """
-
-    def __init__(self, name, memout, timeout, runs, attr):
-        """
-        Initializes a job.
-        
-        Keyword arguments:
-        name    - A unique name for a job
-        timeout - A timeout in seconds for individual benchmark runs
-        runs    - The number of runs per benchmark
-        attr    - A dictionary of arbitrary attributes
-        """
-        self.name = name
-        self.timeout = timeout
-        self.runs = runs
-        self.attr = attr
-        self.memout = memout
-
-    def _toXml(self, out, indent, xmltag, extra):
-        """
-        Helper function to dump a (pretty-printed) XML-representation of a job.
-        
-        Keyword arguments:
-        out     - Output stream to write to
-        indent  - Amount of indentation
-        xmltag  - Tag name for the job
-        extra   - Additional arguments for the job 
-        """
-        out.write(
-            '{1}<{2} name="{0.name}" timeout="{0.timeout}" runs="{0.runs}"{3}'.format(self, indent, xmltag, extra))
-        for key, val in self.attr.items():
-            out.write(' {0}="{1}"'.format(key, val))
-        out.write('/>\n')
-
-    def __hash__(self):
-        """
-        Calculates a hash value for the job using its name. 
-        """
-        return hash(self.name)
-
-    def __cmp__(self, job):
-        """
-        Compares two jobs using their names. 
-        """
-        return cmp(job.name, job.name)
-
-
-class Run(Sortable):
-    """
-    Base class for all runs.
-    
-    Class members:
-    path - Path that holds the target location for start scripts
-    root - directory relative to the location of the run's path.
-    """
-
-    def __init__(self, path):
-        """
-        Initializes a run.
-        
-        Keyword arguments:
-        path - A path that holds the location 
-               where the individual start scripts for the job shall be generated 
-        """
-        self.path = path
-        self.root = os.path.relpath(".", self.path)
-
-
-class SeqRun(Run):
-    """
-    Describes a sequential run.
-    
-    Class members:
-    run      - The number of the run
-    job      - A reference to the job description
-    runspec  - A reference to the run description
-    instance - A reference to the instance to benchmark
-    file     - A relative path to the instance
-    args     - The command line arguments for this run
-    solver   - The solver for this run
-    timeout  - The timeout of this run
-    """
-
-    def __init__(self, path, run, job, runspec, instance):
-        """
-        Initializes a sequential run.
-        
-        Keyword arguments:
-        path     - A path that holds the location 
-                   where the individual start scripts for the job shall be generated
-        run      - The number of the run
-        job      - A reference to the job description
-        runspec  - A reference to the run description
-        instance - A reference to the instance to benchmark
-        """
-        Run.__init__(self, path)
-        self.run = run
-        self.job = job
-        self.runspec = runspec
-        self.instance = instance
-        self.file = os.path.relpath(self.instance.path(), self.path)
-        self.args = self.runspec.setting.cmdline
-        self.solver = self.runspec.system.name + "-" + self.runspec.system.version
-        self.timeout = self.job.timeout
-        self.memlimit = self.job.memout
-        self.instancebase = os.path.basename(instance.path())
-        self.finished = os.path.join(path, ".finished")
-
-
-
-class ScriptGen:
-    """
-    A class providing basic functionality to generate 
-    start scripts for arbitrary jobs and evaluation of results.
-    """
-
-    def __init__(self, job):
-        """
-        Initializes the script generator.
-        
-        Keyword arguments:
-        seqJob - A reference to the associated job.
-        """
-        self.skip = False
-        self.job = job
-        self.startfiles = []
-
-    def setSkip(self, skip):
-        self.skip = skip
-
-    def _path(self, runspec, instance, run):
-        """
-        Returns the relative path to the start script location.
-        
-        Keyword arguments:
-        runspec  - The run specification for the start script
-        instance - The benchmark instance for the start script
-        run      - The number of the run for the start script
-        """
-        return os.path.join(runspec.path(), instance.classname.name, instance.instance, "run%d" % run)
-
-    def addToScript(self, runspec, instance):
-        """
-        Creates a new start script for the given instance.
-        
-        Keyword arguments:
-        runspec  - The run specification for the start script
-        instance - The benchmark instance for the start script
-        """
-        skip = self.skip
-        for run in range(1, self.job.runs + 1):
-            path = self._path(runspec, instance, run)
-            tools.mkdir_p(path)
-            startpath = os.path.join(path, "start.sh")
-            finish = os.path.join(path, ".finished")
-            if skip and os.path.isfile(finish):
-                continue
-            template = open(runspec.system.config.template).read()
-            startfile = open(startpath, "w")
-            startfile.write(template.format(run=SeqRun(path, run, self.job, runspec, instance)))
-
-            startfile.close()
-            self.startfiles.append((runspec, path, "start.sh"))
-            tools.setExecutable(startpath)
-
-    def evalResults(self, out, indent, runspec, instance):
-        """
-        Parses the results of a given benchmark instance and outputs them as XML.
-        
-        Keyword arguments:
-        out     - Output stream to write to
-        indent  - Amount of indentation
-        runspec - The run specification of the benchmark
-        runspec - The benchmark instance
-        """
-        for run in range(1, self.job.runs + 1):
-            out.write('{0}<run number="{1}">\n'.format(indent, run))
-            measure_function = tools.import_function(runspec.system.measures)
-            result = measure_function(self._path(runspec, instance, run), runspec,
-                                                                            instance)
-            for key, valtype, val in sorted(result):
-                out.write('{0}<measure name="{1}" type="{2}" val="{3}"/>\n'.format(indent + "\t", key, valtype, val))
-            out.write('{0}</run>\n'.format(indent))
-
-
-class SeqScriptGen(ScriptGen):
-    """
-    A class that generates and evaluates start scripts for sequential runs.
-    """
-
-    def __init__(self, seqJob):
-        """
-        Initializes the script generator.
-        
-        Keyword arguments:
-        seqJob - A reference to the associated sequential job.
-        """
-        ScriptGen.__init__(self, seqJob)
-
-    def genStartScript(self, path):
-        """
-        Generates a start script that can be used to start all scripts 
-        generated using addToScript().
-        
-        Keyword arguments:
-        path - The target location for the script
-        """
-        tools.mkdir_p(path)
-        startfile = open(os.path.join(path, "start.py"), 'w')
-        queue = ""
-        comma = False
-        for (_, instpath, instname) in self.startfiles:
-            relpath = os.path.relpath(instpath, path)
-            if comma:
-                queue += ","
-            else:
-                comma = True
-            queue += repr(os.path.join(relpath, instname))
-        startfile.write("""\
-#!/usr/bin/python -u
-
-import optparse
-import threading
-import subprocess
-import os
-import sys
-import signal
-import time
-
-queue = [{0}]
-
-class Main:
-    def __init__(self):
-        self.running  = set()
-        self.cores    = set()
-        self.started  = 0
-        self.total    = None
-        self.finished = threading.Condition()
-        self.coreLock = threading.Lock()
-        c = 0
-        while len(self.cores) < {1}:
-            self.cores.add(c)
-            c += 1
-    
-    def finish(self, thread):
-        self.finished.acquire()
-        self.running.remove(thread)
-        with self.coreLock:
-            self.cores.add(thread.core)
-        self.finished.notify()
-        self.finished.release()
-   
-    def start(self, cmd):
-        core     = 0
-        with self.coreLock:
-            core = self.cores.pop()
-        thread = Run(cmd, self, core)
-        self.started += 1
-        self.running.add(thread)
-        print("({{0}}/{{1}}/{{2}}/{{4}}) {{3}}".format(len(self.running), self.started, self.total, cmd, core))
-        thread.start()
-    
-    def run(self, queue):
-        signal.signal(signal.SIGTERM, self.exit)
-        signal.signal(signal.SIGINT, self.exit)
-        signal.signal(signal.SIGHUP, signal.SIG_IGN)
-        self.finished.acquire()
-        self.total = len(queue)
-        for cmd in queue:
-            while len(self.running) >= {1}:
-                self.finished.wait()
-            self.start(cmd)
-        while len(self.running) != 0:
-            self.finished.wait()
-        self.finished.release()
-
-    def exit(self, *args):
-        signal.signal(signal.SIGINT, signal.SIG_IGN)
-        signal.signal(signal.SIGTERM, signal.SIG_IGN)
-        print "WARNING: it is not guaranteed that all processes will be terminated!"
-        print "sending sigterm ..."
-        os.killpg(os.getpgid(0), signal.SIGTERM)
-        print "waiting 10s..."
-        time.sleep(10)
-        print "sending sigkill ..."
-        os.killpg(os.getpgid(0), signal.SIGKILL)
-
-class Run(threading.Thread):
-    def __init__(self, cmd, main, core):
-        threading.Thread.__init__(self)
-        self.cmd  = cmd
-        self.main = main
-        self.core = core
-        self.proc = None
-    
-    def run(self):
-        path, script = os.path.split(self.cmd)
-        self.proc = subprocess.Popen(["bash", script, str(self.core)], cwd=path)
-        self.proc.wait()
-        self.main.finish(self)
-
-def gui():
-    import Tkinter
-    class App:
-        def __init__(self):
-            root    = Tkinter.Tk()
-            frame   = Tkinter.Frame(root)
-            scrollx = Tkinter.Scrollbar(frame, orient=Tkinter.HORIZONTAL)
-            scrolly = Tkinter.Scrollbar(frame)
-            list    = Tkinter.Listbox(frame, selectmode=Tkinter.MULTIPLE)
-            
-            for script in queue:
-                list.insert(Tkinter.END, script)
-            
-            scrolly.config(command=list.yview)
-            scrollx.config(command=list.xview)
-            list.config(yscrollcommand=scrolly.set)
-            list.config(xscrollcommand=scrollx.set)
-                
-            scrolly.pack(side=Tkinter.RIGHT, fill=Tkinter.Y)
-            scrollx.pack(side=Tkinter.BOTTOM, fill=Tkinter.X)
-            list.pack(fill=Tkinter.BOTH, expand=1)
-            
-            button = Tkinter.Button(root, text='Run', command=self.pressed)
-            
-            frame.pack(fill=Tkinter.BOTH, expand=1)
-            button.pack(side=Tkinter.BOTTOM, fill=Tkinter.X)
-
-            self.root  = root
-            self.list  = list
-            self.run   = False
-            self.queue = [] 
-        
-        def pressed(self):
-            sel = self.list.curselection()
-            for index in sel:
-                global queue
-                self.queue.append(queue[int(index)])
-            self.root.destroy()
-
-        def start(self):
-            self.root.mainloop()
-            return self.queue
-
-    global queue
-    queue.sort()
-    queue = App().start()
-
-if __name__ == '__main__':
-    usage  = "usage: %prog [options] <runscript>"
-    parser = optparse.OptionParser(usage=usage)
-    parser.add_option("-g", "--gui", action="store_true", dest="gui", default=False, help="start gui to selectively start benchmarks") 
-
-    opts, args = parser.parse_args(sys.argv[1:])
-    if len(args) > 0: parser.error("no arguments expected")
-    
-    os.chdir(os.path.dirname(sys.argv[0]))
-    if opts.gui: gui()
-
-    m = Main()
-    m.run(queue)
-""".format(queue, self.job.parallel))
-        startfile.close()
-        tools.setExecutable(os.path.join(path, "start.py"))
-
-
-class PbsScriptGen(ScriptGen):
-    """
-    A class that generates and evaluates start scripts for pbs runs.
-    """
-
-    class PbsScript:
-        def __init__(self, runspec, path, queue):
-            self.runspec = runspec
-            self.path = path
-            self.queue = queue
-            self.num = 0
-            self.time = None
-            self.startscripts = None
-            self.next()
-
-        def write(self):
-            if self.num > 0:
-                self.num = 0
-                template = open(self.runspec[2], "r").read()
-                script = os.path.join(self.path, "start{0:04}.pbs".format(len(self.queue)))
-                open(script, "w").write(
-                    template.format(walltime=tools.pbsTime(self.runspec[3]), nodes=self.runspec[1], ppn=self.runspec[0],
-                                    jobs=self.startscripts))
-                self.queue.append(script)
-
-        def next(self):
-            self.write()
-            self.startscripts = ""
-            self.num = 0
-            self.time = 0
-
-        def append(self, startfile):
-            self.num += 1
-            self.startscripts += startfile + "\n"
-
-    def __init__(self, seqJob):
-        """
-        Initializes the script generator.
-        
-        Keyword arguments:
-        seqJob - A reference to the associated sequential job.
-        """
-        ScriptGen.__init__(self, seqJob)
-
-    def genStartScript(self, path):
-        """
-        Generates a start script that can be used to start all scripts 
-        generated using addToScript().
-        
-        Keyword arguments:
-        path - The target location for the script
-        """
-        tools.mkdir_p(path)
-        startfile = open(os.path.join(path, "start.sh"), 'w')
-        queue = []
-        pbsScripts = {}
-        for (runspec, instpath, instname) in self.startfiles:
-            #TODO: move to xml file
-            if instname.startswith('move.sh') or instname.startswith('results.txt'):
-                continue
-            relpath = os.path.relpath(instpath, path)
-            jobScript = os.path.join(relpath, instname)
-            pbsKey = (
-                runspec.setting.ppn, runspec.setting.procs, runspec.setting.pbstemplate, runspec.project.job.walltime)
-
-            if not pbsKey in pbsScripts:
-                pbsScript = PbsScriptGen.PbsScript(pbsKey, path, queue)
-                pbsScripts[pbsKey] = pbsScript
-            else:
-                pbsScript = pbsScripts[pbsKey]
-
-            if self.job.script_mode == "multi":
-                if pbsScript.num > 0: pbsScript.next()
-                pbsScript.append(jobScript)
-            elif self.job.script_mode == "timeout":
-                if pbsScript.time + runspec.project.job.timeout + 300 >= runspec.project.job.walltime:
-                    pbsScript.next()
-                pbsScript.time += runspec.project.job.timeout + 300
-                pbsScript.append(jobScript)
-
-        for pbsScript in pbsScripts.values(): pbsScript.write()
-
-        startfile.write("""#!/bin/bash\n\ncd "$(dirname $0)"\n""" + "\n".join(
-            ['qsub "{0}"'.format(os.path.basename(x)) for x in queue]))
-        startfile.close()
-        tools.setExecutable(os.path.join(path, "start.sh"))
-
-
-class CondorScriptGen(ScriptGen):
-    """
-    A class that generates and evaluates start scripts for pbs runs.
-    """
-
-    class CondorScript:
-        def __init__(self, runspec, path, queue):
-            self.runspec = runspec
-            self.path = path
-            self.queue = queue
-            self.num = 0
-            self.time = None
-            self.startscripts = None
-            self.next()
-
-        def write(self):
-            if self.num > 0:
-                self.num = 0
-                template = open(self.runspec[2], "r").read()
-                script = os.path.join(self.path, "start{0:04}.pbs".format(len(self.queue)))
-                open(script, "w").write(
-                    template.format(walltime=tools.pbsTime(self.runspec[3]), nodes=self.runspec[1], ppn=self.runspec[0],
-                                    jobs=self.startscripts))
-                self.queue.append(script)
-
-        def next(self):
-            self.write()
-            self.startscripts = ""
-            self.num = 0
-            self.time = 0
-
-        def append(self, startfile):
-            self.num += 1
-            self.startscripts += startfile + "\n"
-
-    def __init__(self, seqJob):
-        """
-        Initializes the script generator.
-
-        Keyword arguments:
-        seqJob - A reference to the associated sequential job.
-        """
-        ScriptGen.__init__(self, seqJob)
-
-    def addToScript(self, runspec, instance):
-        skip = self.skip
-        for run in range(1, self.job.runs + 1):
-            path = self._path(runspec, instance, run)
-            tools.mkdir_p(path)
-            startpath = os.path.join(path, "start.py")
-            finish = os.path.join(path, ".finished")
-            if skip and os.path.isfile(finish):
-                continue
-            seq = SeqRun(path, run, self.job, runspec, instance)
-            with open(startpath, "w") as startfile:
-                with open(runspec.system.config.template) as templatefile:
-                    template = templatefile.read()
-                run = dotdict(seq.__dict__)
-                run.instancebase = os.path.basename(run.instance.path())
-                run.root = self.job.basedir.strip()
-                run.instance = run.instance.path()
-                run.memlimit = self.job.memout
-                run.timelimit = self.job.timeout
-                run.finished = finish
-                startfile.write(template.format(run=run))
-            os.chmod(startpath, 0o770)
-            self.startfiles.append((runspec, path, "start.py"))
-            tools.setExecutable(startpath)
-
-        return
-
-    def genStartScript(self, path):
-        """
-        Generates a condor submit script that can be used to start all scripts
-        generated using addToScript().
-
-        Keyword arguments:
-        path - The target location for the script
-        """
-        tools.mkdir_p(path)
-
-        instances = []
-        initialdir = os.path.join(self.job.basedir.strip(), 'output')
-        for (runspec, instpath, instname) in self.startfiles:
-            instance = os.path.join(self.job.basedir.strip(), instpath, instname)
-            outprefix = os.path.join(self.job.basedir.strip(), instpath)
-            instances.append((instance, outprefix))
-
-        for i, chunk in enumerate(chunks(instances, 5000)):
-            startpath = os.path.join(self.job.basedir.strip(), path, "condor_%s.submit" %i)
-            with open(startpath, "w") as condorsubmitfile:
-                with open(self.job.condortemplate) as template:
-                    t = Template(template.read())
-                    condorsubmitfile.write(
-                        t.render(instances=chunk, timeout=self.job.timeout,
-                                 memout=self.job.memout, initialdir=initialdir))
-        #tools.setExecutable(os.path.join(path, "condor_0.submit"))
-
-
-class SeqJob(Job):
-    """
-    Describes a sequential job.
-    """
-
-    def __init__(self, name, memout, timeout, runs, parallel, attr):
-        """
-        Initializes a sequential job description.  
-        
-        Keyword arguments:
-        name     - A unique name for a job
-        timeout  - A timeout in seconds for individual benchmark runs
-        runs     - The number of runs per benchmark
-        parallel - The number of runs that can be started in parallel
-        attr     - A dictionary of arbitrary attributes
-        """
-        Job.__init__(self, name, memout, timeout, runs, attr)
-        self.parallel = parallel
-
-    def scriptGen(self):
-        """
-        Returns a class that can generate start scripts and evaluate benchmark results.
-        (see SeqScriptGen) 
-        """
-        return SeqScriptGen(self)
-
-    def toXml(self, out, indent):
-        """
-        Dump a (pretty-printed) XML-representation of the sequential job.
-        
-        Keyword arguments:
-        out     - Output stream to write to
-        indent  - Amount of indentation
-        """
-        extra = ' parallel="{0.parallel}"'.format(self)
-        Job._toXml(self, out, indent, "seqjob", extra)
-
-
-class PbsJob(Job):
-    """
-    Describes a pbs job.
-    """
-
-    def __init__(self, name, memout, timeout, runs, script_mode, walltime, attr):
-        """
-        Initializes a parallel job description.  
-        
-        Keyword arguments:
-        name        - A unique name for a job
-        timeout     - A timeout in seconds for individual benchmark runs
-        runs        - The number of runs per benchmark
-        script_mode - Specifies the script generation mode
-        walltime    - The walltime for a job submitted via PBS 
-        attr        - A dictionary of arbitrary attributes
-        """
-        Job.__init__(self, name, memout, timeout, runs, attr)
-        self.script_mode = script_mode
-        self.walltime = walltime
-
-    def toXml(self, out, indent):
-        """
-        Dump a (pretty-printed) XML-representation of the parallel job.
-        
-        Keyword arguments:
-        out     - Output stream to write to
-        indent  - Amount of indentation
-        """
-        extra = ' script_mode="{0.script_mode}" walltime="{0.walltime}"'.format(self)
-        Job._toXml(self, out, indent, "pbsjob", extra)
-
-    def scriptGen(self):
-        """
-        Returns a class that can generate start scripts and evaluate benchmark results.
-        (see SeqScriptGen) 
-        """
-        return PbsScriptGen(self)
-
-
-class CondorJob(Job):
-    """
-    Describes a condor job.
-    """
-
-    def __init__(self, name, memout, timeout, runs, script_mode, walltime, condortemplate, basedir, attr):
-        """
-        Initializes a parallel job description.
-
-        Keyword arguments:
-        name        - A unique name for a job
-        timeout     - A timeout in seconds for individual benchmark runs
-        runs        - The number of runs per benchmark
-        script_mode - Specifies the script generation mode
-        walltime    - The walltime for a job submitted via PBS
-        attr        - A dictionary of arbitrary attributes
-        """
-        Job.__init__(self, name, memout, timeout, runs, attr)
-        self.script_mode = script_mode
-        self.walltime = walltime
-        self.condortemplate = condortemplate
-        self.basedir = basedir
-
-    def toXml(self, out, indent):
-        """
-        Dump a (pretty-printed) XML-representation of the parallel job.
-
-        Keyword arguments:
-        out     - Output stream to write to
-        indent  - Amount of indentation
-        """
-        extra = ' script_mode="{0.script_mode}" walltime="{0.walltime}"'.format(self)
-        Job._toXml(self, out, indent, "pbsjob", extra)
-
-    def scriptGen(self):
-        """
-        Returns a class that can generate start scripts and evaluate benchmark results.
-        (see SeqScriptGen)
-        """
-        return CondorScriptGen(self)
-
 
 class Config(Sortable):
     """
